+++ src/main/java/user/LoginResponse.java
@@ -1,3 +1,5 @@
+package user;
+
 public class LoginResponse {
     private String accessToken;
     private String refreshToken;
