+++ src/main/java/user/User.java
@@ -1,3 +1,5 @@
+package user;
+
 public class User {
     private String email;
     private String password;
