+++ src/main/java/user/UserCreds.java
@@ -1,3 +1,5 @@
+package user;
+
 public class UserCreds {
     private String email;
     private String password;
