+++ src/main/java/user/UserResponse.java
@@ -1,3 +1,5 @@
+package user;
+
 public class UserResponse {
     private String name;
     private String email;
