+++ src/test/java/order/CreateOrderWithAuthTest.java
@@ -1,19 +1,23 @@
+package order;
+
 import com.github.javafaker.Faker;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
+import order.*;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import user.*;
 
 import java.util.ArrayList;
 
-
 import static constant.Api.BASE_URL;
 import static org.hamcrest.core.IsEqual.equalTo;
+import static org.hamcrest.core.IsNull.notNullValue;
 
-public class CreateOrderWithoutAuthTest {
+public class CreateOrderWithAuthTest {
 
 
     private String token;
@@ -34,37 +38,56 @@ public class CreateOrderWithoutAuthTest {
         token = loginResponse.body().as(LoginResponse.class).getAccessToken();
     }
 
-    @DisplayName("Создание заказа с одним ингредиентом под неавторизованным пользователем")
+    @DisplayName("Создание заказа с одним ингредиентом под авторизованным пользователем")
     @Test
     public void createOrderOneIngredient() {
         Order order = new Order();
         int numberIngredient = faker.number().numberBetween(0, allIngredients.size() - 1);
         IngredientsForOrder ingredientsForOrder = new IngredientsForOrder();
         ingredientsForOrder.getIngredients().add(allIngredients.get(numberIngredient).getId());
-        Response createOrder = order.createOrderWithoutAuth(ingredientsForOrder);
+        Response createOrder = order.createOrderWithAuth(token, ingredientsForOrder);
         createOrder.then()
-                .statusCode(401)
+                .statusCode(200)
                 .and()
-                .body("success", equalTo(false))
+                .body("success", equalTo(true))
                 .and()
-                .body("message", equalTo("You should be authorised"));
+                .body("name", notNullValue());
     }
 
-    @DisplayName("Создание заказа без ингредиентов под неавторизованным пользователем")
+    @DisplayName("Создание заказа с несколькими ингредиентами под авторизованным пользователем")
+    @Test
+    public void createOrderFewIngredients() {
+        Order order = new Order();
+        IngredientsForOrder ingredientsForOrder = new IngredientsForOrder();
+        int countIngredient = faker.number().numberBetween(2, allIngredients.size() - 2);
+        for (int i = 0; i <= countIngredient; i++) {
+            int numberIngredient = faker.number().numberBetween(0, allIngredients.size() - 1);
+            ingredientsForOrder.getIngredients().add(allIngredients.get(numberIngredient).getId());
+        }
+        Response createOrder = order.createOrderWithAuth(token, ingredientsForOrder);
+        createOrder.then()
+                .body("success", equalTo(true))
+                .and()
+                .statusCode(200)
+                .and()
+                .body("name", notNullValue());
+    }
+
+    @DisplayName("Создание заказа без ингредиентов под авторизованным пользователем")
     @Test
     public void createOrderWithoutIngredients() {
         Order order = new Order();
         IngredientsForOrder ingredientsForOrder = new IngredientsForOrder();
-        Response createOrder = order.createOrderWithoutAuth(ingredientsForOrder);
+        Response createOrder = order.createOrderWithAuth(token, ingredientsForOrder);
         createOrder.then()
                 .body("success", equalTo(false))
                 .and()
-                .statusCode(401)
+                .statusCode(400)
                 .and()
-                .body("message", equalTo("You should be authorised"));
+                .body("message", equalTo("Ingredient ids must be provided"));
     }
 
-    @DisplayName("Создание заказа с неправильным хэшем ингредиента под неавторизованным пользователем")
+    @DisplayName("Создание заказа с неправильным хэшем ингредиента под авторизованным пользователем")
     @Test
     public void createOrderWithWrongIngredients() {
         Order order = new Order();
@@ -75,13 +98,9 @@ public class CreateOrderWithoutAuthTest {
             ingredientsForOrder.getIngredients().add(allIngredients.get(numberIngredient).getId());
         }
         ingredientsForOrder.getIngredients().set(0, ingredientsForOrder.getIngredients().get(0) + "test");
-        Response createOrder = order.createOrderWithoutAuth(ingredientsForOrder);
+        Response createOrder = order.createOrderWithAuth(token, ingredientsForOrder);
         createOrder.then()
-                .statusCode(401)
-                .and()
-                .body("message", equalTo("You should be authorized"))
-                .and()
-                .body("success", equalTo(false));
+                .statusCode(500);
     }
 
     @After
