+++ src/test/java/order/CreateOrderWithoutAuthTest.java
@@ -1,19 +1,23 @@
+package order;
+
 import com.github.javafaker.Faker;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
+import order.*;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import user.*;
 
 import java.util.ArrayList;
 
+
 import static constant.Api.BASE_URL;
 import static org.hamcrest.core.IsEqual.equalTo;
-import static org.hamcrest.core.IsNull.notNullValue;
 
-public class CreateOrderWithAuthTest {
+public class CreateOrderWithoutAuthTest {
 
 
     private String token;
@@ -34,56 +38,37 @@ public class CreateOrderWithAuthTest {
         token = loginResponse.body().as(LoginResponse.class).getAccessToken();
     }
 
-    @DisplayName("Создание заказа с одним ингредиентом под авторизованным пользователем")
+    @DisplayName("Создание заказа с одним ингредиентом под неавторизованным пользователем")
     @Test
     public void createOrderOneIngredient() {
         Order order = new Order();
         int numberIngredient = faker.number().numberBetween(0, allIngredients.size() - 1);
         IngredientsForOrder ingredientsForOrder = new IngredientsForOrder();
         ingredientsForOrder.getIngredients().add(allIngredients.get(numberIngredient).getId());
-        Response createOrder = order.createOrderWithAuth(token, ingredientsForOrder);
+        Response createOrder = order.createOrderWithoutAuth(ingredientsForOrder);
         createOrder.then()
-                .statusCode(200)
+                .statusCode(401)
                 .and()
-                .body("success", equalTo(true))
+                .body("success", equalTo(false))
                 .and()
-                .body("name", notNullValue());
+                .body("message", equalTo("You should be authorised"));
     }
 
-    @DisplayName("Создание заказа с несколькими ингредиентами под авторизованным пользователем")
-    @Test
-    public void createOrderFewIngredients() {
-        Order order = new Order();
-        IngredientsForOrder ingredientsForOrder = new IngredientsForOrder();
-        int countIngredient = faker.number().numberBetween(2, allIngredients.size() - 2);
-        for (int i = 0; i <= countIngredient; i++) {
-            int numberIngredient = faker.number().numberBetween(0, allIngredients.size() - 1);
-            ingredientsForOrder.getIngredients().add(allIngredients.get(numberIngredient).getId());
-        }
-        Response createOrder = order.createOrderWithAuth(token, ingredientsForOrder);
-        createOrder.then()
-                .body("success", equalTo(true))
-                .and()
-                .statusCode(200)
-                .and()
-                .body("name", notNullValue());
-    }
-
-    @DisplayName("Создание заказа без ингредиентов под авторизованным пользователем")
+    @DisplayName("Создание заказа без ингредиентов под неавторизованным пользователем")
     @Test
     public void createOrderWithoutIngredients() {
         Order order = new Order();
         IngredientsForOrder ingredientsForOrder = new IngredientsForOrder();
-        Response createOrder = order.createOrderWithAuth(token, ingredientsForOrder);
+        Response createOrder = order.createOrderWithoutAuth(ingredientsForOrder);
         createOrder.then()
                 .body("success", equalTo(false))
                 .and()
-                .statusCode(400)
+                .statusCode(401)
                 .and()
-                .body("message", equalTo("Ingredient ids must be provided"));
+                .body("message", equalTo("You should be authorised"));
     }
 
-    @DisplayName("Создание заказа с неправильным хэшем ингредиента под авторизованным пользователем")
+    @DisplayName("Создание заказа с неправильным хэшем ингредиента под неавторизованным пользователем")
     @Test
     public void createOrderWithWrongIngredients() {
         Order order = new Order();
@@ -94,9 +79,13 @@ public class CreateOrderWithAuthTest {
             ingredientsForOrder.getIngredients().add(allIngredients.get(numberIngredient).getId());
         }
         ingredientsForOrder.getIngredients().set(0, ingredientsForOrder.getIngredients().get(0) + "test");
-        Response createOrder = order.createOrderWithAuth(token, ingredientsForOrder);
+        Response createOrder = order.createOrderWithoutAuth(ingredientsForOrder);
         createOrder.then()
-                .statusCode(500);
+                .statusCode(401)
+                .and()
+                .body("message", equalTo("You should be authorized"))
+                .and()
+                .body("success", equalTo(false));
     }
 
     @After
