+++ src/test/java/order/GetOrderUserWithoutAuthTest.java
@@ -1,11 +1,15 @@
+package order;
+
 import com.github.javafaker.Faker;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
+import order.*;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import user.*;
 
 import java.util.ArrayList;
