+++ src/test/java/user/UserUpdateTest.java
@@ -1,3 +1,5 @@
+package user;
+
 import com.github.javafaker.Faker;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.RestAssured;
@@ -6,6 +8,7 @@ import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import user.*;
 
 import static constant.Api.BASE_URL;
 import static org.hamcrest.core.IsEqual.equalTo;
